+++ Package.swift
@@ -1,5 +1,5 @@
 // swift-tools-version: 6.0
-// SixLayerFramework v7.0.2 - Hints File Presentation Properties Support
+// SixLayerFramework v7.2.0 - Configurable Photo Sources for OCR Scanner
 // The swift-tools-version declares the minimum version of Swift required to build this package.
 
 import PackageDescription